# packages/api-contracts/openapi/wishpool.yaml
openapi: 3.1.0
info:
  title: WishPool API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    WishPool core API contract for mobile apps, parent web, admin web, and backend services.
    All write endpoints are command-oriented and should be called with an Idempotency-Key header.
servers:
  - url: http://localhost:8080
    description: Local core-api
security:
  - bearerAuth: []
...
paths:
...
  /families/{familyId}/pairing-sessions:
    post:
      tags: [Pairing]
      operationId: createPairingSession
      summary: Create a child device pairing session.
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePairingSessionRequest'
      responses:
        '201':
          description: Pairing session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairingSessionCreated'
        default:
          $ref: '#/components/responses/Error'
  /pairing/consume:
    post:
      tags: [Pairing]
      security: []
      operationId: consumePairingCode
      summary: Pair a child device with a family.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumePairingCodeRequest'
      responses:
        '200':
          description: Device paired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenPair'
        default:
          $ref: '#/components/responses/Error'
...
  /reviews/pending:
    get:
      tags: [Reviews]
      operationId: listPendingReviews
      summary: List pending reviews.
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pending review cards.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingReviewCard'
        default:
          $ref: '#/components/responses/Error'
  /reviews:
    post:
      tags: [Reviews]
      operationId: reviewSubmission
      summary: Approve or reject a submission.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSubmissionRequest'
      responses:
        '201':
          description: Review created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        default:
          $ref: '#/components/responses/Error'
  /reviews/{submissionId}/detail:
    get:
      tags: [Reviews]
      operationId: getReviewDetail
      summary: Get a review detail view for a submission.
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
        default:
          $ref: '#/components/responses/Error'
  /reviews/{reviewId}/revoke:
    post:
      tags: [Reviews]
      operationId: revokeReview
      summary: Revoke a review and create adjustment records when needed.
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeReviewRequest'
      responses:
        '200':
          description: Review revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        default:
          $ref: '#/components/responses/Error'
...
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: List the current user's inbox notifications.
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, sent, failed, read]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Inbox notifications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        default:
          $ref: '#/components/responses/Error'
  /notifications/read:
    post:
      tags: [Notifications]
      operationId: markNotificationsRead
      summary: Mark inbox notifications as read.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationReadRequest'
      responses:
        '200':
          description: Updated notifications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        default:
          $ref: '#/components/responses/Error'
  /notification-preferences:
    get:
      tags: [Notifications]
      operationId: listNotificationPreferences
      summary: List notification preferences for the current user.
      parameters:
        - name: familyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification preferences.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationPreference'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Notifications]
      operationId: updateNotificationPreference
      summary: Upsert one notification preference for the current user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationPreferenceRequest'
      responses:
        '200':
          description: Notification preference.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreference'
        default:
          $ref: '#/components/responses/Error'
  /devices/push-token:
    post:
      tags: [Notifications]
      operationId: registerPushToken
      summary: Register or update a push token for the current device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPushTokenRequest'
      responses:
        '200':
          description: Registered push-capable device.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDevice'
        default:
          $ref: '#/components/responses/Error'
  /privacy/export:
    post:
      tags: [Privacy]
      operationId: requestDataExport
      summary: Request family data export.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequestCreate'
      responses:
        '202':
          description: Export requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequest'
        default:
          $ref: '#/components/responses/Error'
  /privacy/delete:
    post:
      tags: [Privacy]
      operationId: requestFamilyDeletion
      summary: Request family deletion.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequestCreate'
      responses:
        '202':
          description: Deletion requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequest'
        default:
          $ref: '#/components/responses/Error'
  /internal/notifications:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: createNotification
      summary: Create or deduplicate an internal notification event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationEventRequest'
      responses:
        '200':
          description: Notification event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEvent'
        default:
          $ref: '#/components/responses/Error'
  /internal/notifications/claim:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: claimNotificationEvents
      summary: Claim pending notification events for dispatch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimNotificationEventsRequest'
      responses:
        '200':
          description: Claimed notifications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimNotificationEventsResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/notifications/{notificationId}/dispatch-result:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: markNotificationDispatchResult
      summary: Mark notification dispatch result.
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationDispatchResultRequest'
      responses:
        '200':
          description: Notification event.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationEvent'
        default:
          $ref: '#/components/responses/Error'
...
  /internal/outbox/events/claim:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: claimOutboxEvents
      summary: Claim unpublished outbox events for an internal publisher.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimOutboxEventsRequest'
      responses:
        '200':
          description: Claimed outbox events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxClaimResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/outbox/events/{eventId}/published:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: markOutboxEventPublished
      summary: Mark an outbox event as published.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Outbox event publication acknowledged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxAckResponse'
        default:
          $ref: '#/components/responses/Error'
  /internal/outbox/events/{eventId}/retry:
    post:
      tags: [Internal]
      security:
        - internalToken: []
      operationId: retryOutboxEvent
      summary: Release an outbox event lease and schedule retry.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryOutboxEventRequest'
      responses:
        '200':
          description: Outbox event retry scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxEvent'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
...
    AuthTokenPair:
      type: object
      required: [accessToken, refreshToken, expiresInSec, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSec:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        primaryFamilyId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
...
    User:
      type: object
      required: [id, status, displayName]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, suspended, deleted]
        displayName:
          type: string
        avatarUrl:
          anyOf:
            - type: string
            - type: 'null'
    DeviceRegistration:
      type: object
      required: [platform]
      properties:
        platform:
          type: string
          enum: [ios, android, ipad_os, web, admin_web]
        deviceName:
          type: string
        pushProvider:
          type: string
          enum: [apns, fcm, huawei, xiaomi, oppo, vivo]
        pushToken:
          type: string
...
    CreatePairingSessionRequest:
      type: object
      required: [childId]
      properties:
        childId:
          type: string
          format: uuid
    PairingSessionCreated:
      type: object
      required: [sessionId, pairingCode, expiresAt]
      properties:
        sessionId:
          type: string
          format: uuid
        pairingCode:
          type: string
        expiresAt:
          type: string
          format: date-time
    ConsumePairingCodeRequest:
      type: object
      required: [pairingCode, device]
      properties:
        pairingCode:
          type: string
        device:
          $ref: '#/components/schemas/DeviceRegistration'
...
    SubmissionDetail:
      allOf:
        - $ref: '#/components/schemas/Submission'
        - type: object
          properties:
            task:
              $ref: '#/components/schemas/TaskInstance'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaAsset'
            aiPrecheck:
              $ref: '#/components/schemas/AiPrecheck'
            review:
              $ref: '#/components/schemas/Review'
...
    PendingReviewCard:
      type: object
      required: [submission, task]
      properties:
        submission:
          $ref: '#/components/schemas/Submission'
        task:
          $ref: '#/components/schemas/TaskInstance'
        aiSummary:
          anyOf:
            - type: string
            - type: 'null'
        thumbnailMedia:
          $ref: '#/components/schemas/MediaAsset'
    ReviewSubmissionRequest:
      type: object
      required: [submissionId, decision]
      properties:
        submissionId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, needs_revision]
        feedback:
          $ref: '#/components/schemas/FeedbackInput'
...
    Review:
      type: object
      required: [id, submissionId, taskInstanceId, decision, reviewedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        submissionId:
          type: string
          format: uuid
        taskInstanceId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, needs_revision]
        reviewedBy:
          type: string
          format: uuid
        feedback:
          anyOf:
            - $ref: '#/components/schemas/Feedback'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
...
    RevokeReviewRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...
    ClaimOutboxEventsRequest:
      type: object
      required: [limit]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
        leaseSeconds:
          type: integer
          format: int64
          minimum: 5
          maximum: 900
    RetryOutboxEventRequest:
      type: object
      required: [delaySeconds]
      properties:
        delaySeconds:
          type: integer
          format: int64
          minimum: 0
          maximum: 86400
        reason:
          type: string
    OutboxClaimResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/OutboxEvent'
    OutboxEvent:
      type: object
      required: [id, type, aggregateType, aggregateId, payload, availableAt, retryCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        aggregateType:
          type: string
        aggregateId:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
        availableAt:
          type: string
          format: date-time
        leasedUntil:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        retryCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    OutboxAckResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [published]
        publishedAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
...
    NotificationEvent:
      type: object
      required: [id, familyId, recipientUserId, type, title, body, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        recipientUserId:
          type: string
          format: uuid
        recipientDeviceId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        type:
          type: string
          enum: [child_submission_created, ai_precheck_completed, review_completed, wish_fragment_earned, wish_unlocked, wish_redeemed_memory_generated, task_plan_changed]
        title:
          type: string
        body:
          type: string
        relatedResourceType:
          anyOf:
            - type: string
            - type: 'null'
        relatedResourceId:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
        status:
          type: string
          enum: [pending, sent, failed, read, suppressed]
        sentAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        readAt:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationEvent'
        nextCursor:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
    MarkNotificationReadRequest:
      type: object
      required: [notificationIds]
      properties:
        notificationIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
    NotificationPreference:
      type: object
      required: [id, familyId, userId, notificationType, enabled, quietHours, channels, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        notificationType:
          type: string
          enum: [child_submission_created, ai_precheck_completed, review_completed, wish_fragment_earned, wish_unlocked, wish_redeemed_memory_generated, task_plan_changed]
        enabled:
          type: boolean
        quietHours:
          type: object
          additionalProperties: true
        channels:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
    UpdateNotificationPreferenceRequest:
      type: object
      required: [familyId, notificationType, enabled]
      properties:
        familyId:
          type: string
          format: uuid
        notificationType:
          type: string
          enum: [child_submission_created, ai_precheck_completed, review_completed, wish_fragment_earned, wish_unlocked, wish_redeemed_memory_generated, task_plan_changed]
        enabled:
          type: boolean
        quietHours:
          type: object
          additionalProperties: true
        channels:
          type: object
          additionalProperties: true
    RegisterPushTokenRequest:
      type: object
      required: [familyId, platform, pushProvider, pushToken]
      properties:
        familyId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [ios, android, ipad_os, web, admin_web]
        pushProvider:
          type: string
          enum: [apns, fcm, huawei, xiaomi, oppo, vivo]
        pushToken:
          type: string
        deviceName:
          type: string
    NotificationDevice:
      type: object
      required: [id, platform]
      properties:
        id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [ios, android, ipad_os, web, admin_web]
        deviceName:
          anyOf:
            - type: string
            - type: 'null'
        pushProvider:
          anyOf:
            - type: string
              enum: [apns, fcm, huawei, xiaomi, oppo, vivo]
            - type: 'null'
        pushToken:
          anyOf:
            - type: string
            - type: 'null'
    CreateNotificationEventRequest:
      type: object
      required: [familyId, recipientUserId, type, title, body]
      properties:
        familyId:
          type: string
          format: uuid
        recipientUserId:
          type: string
          format: uuid
        recipientDeviceId:
          type: string
          format: uuid
        type:
          type: string
          enum: [child_submission_created, ai_precheck_completed, review_completed, wish_fragment_earned, wish_unlocked, wish_redeemed_memory_generated, task_plan_changed]
        title:
          type: string
        body:
          type: string
        relatedResourceType:
          type: string
        relatedResourceId:
          type: string
          format: uuid
        dedupeKey:
          type: string
    ClaimNotificationEventsRequest:
      type: object
      required: [limit]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
    ClaimNotificationEventsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationDispatchItem'
...
    NotificationDispatchResultRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [sent, failed, suppressed]
        providerMessageId:
          type: string
        errorMessage:
          type: string
...
    PrivacyRequestCreate:
      type: object
      required: [familyId, requestType, confirmationText]
      properties:
        familyId:
          type: string
          format: uuid
        requestType:
          type: string
          enum: [export, delete]
        confirmationText:
          type: string
        reason:
          type: string
    PrivacyRequest:
      type: object
      required: [id, familyId, requestType, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        requestType:
          type: string
          enum: [export, delete]
        status:
          type: string
          enum: [requested, verifying, locking_family, deleting_records, deleting_objects, verifying_deletion, completed, failed_needs_attention]
        exportMedia:
          $ref: '#/components/schemas/MediaAsset'
        createdAt:
          type: string
          format: date-time